service-principal/requirements.txt
pyodbc==5.1.0
azure.identity